pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "libraryreach"
version = "0.1.0"
description = "Library accessibility and outreach planning (baseline transit stop density)."
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "fastapi>=0.110",
  "numpy>=1.24",
  "pandas>=2.0",
  "pydantic>=2.0",
  "PyYAML>=6.0",
  "requests>=2.31",
  "scipy>=1.10",
  "uvicorn>=0.24",
]

[project.optional-dependencies]
dev = ["pytest>=7.0"]

[project.scripts]
libraryreach = "libraryreach.cli:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
libraryreach = ["web/static/*"]
